Revert "Dialog: move files, add a story (#3019)"
This reverts commit 2d5ed00c35f2e7410a4440c052c5738cc9e8b074.
