+++ docs/content/Dialog.mdx
@@ -3,7 +3,7 @@ title: Dialog
 status: Alpha
 ---
 
-import data from '../../src/Dialog/Dialog.docs.json'
+import data from '../../src/Dialog.docs.json'
 
 The dialog component is used for all modals. It renders on top of the rest of the app with an overlay.
