--- src/Dialog/index.ts
@@ -1,1 +1,0 @@
-export {default, DialogProps, DialogHeaderProps} from './Dialog'
