+++ src/__tests__/storybook.test.tsx
@@ -21,7 +21,6 @@ const allowlist = [
   'CounterLabel',
   'DataTable',
   'Details',
-  'Dialog',
   'Dialog2',
   'Flash',
   'Heading',
